# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/reports/dashboard": {
      "get": {
        "operationId": "Reports_dashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Reports"
        ],
        "summary": "Retrieve reports",
        "description": "Retrieve operation in the Reports module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/reports/utilization": {
      "get": {
        "operationId": "Reports_utilization",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "2026-08-01T00:00:00.000Z"
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "2026-08-17T23:59:59.999Z"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Reports"
        ],
        "summary": "Retrieve reports",
        "description": "Retrieve operation in the Reports module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/reports/vehicles/{id}/route": {
      "get": {
        "operationId": "Reports_route",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "2026-08-01T00:00:00.000Z"
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "2026-08-17T23:59:59.999Z"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Reports"
        ],
        "summary": "Retrieve reports",
        "description": "Retrieve operation in the Reports module. Organization scope is enforced from the authenticated session."
      }
    }
  },
  "info": {
    "title": "ORBIT Fleet API",
    "description": "Complete fleet operations API: authentication, vehicles, TCP-normalized telemetry, drivers, geofences, alerts, maintenance, fuel, tasks, device commands, cameras, configuration, users, audit and reports.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the access token returned by POST /v1/auth/login"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "HttpOnly browser session cookie"
      }
    },
...
  }
}